New function PluginClient::get_picker_colors
Returns color picker colors at once
Should use colors of plugin->edl what may differ from master_edl
Replaces get_red(), get_green(), get_blue()

+++ plugins/chromakey/chromakey.h
@@ -57,9 +57,9 @@ public:
 		ptstime current_pts);
 	int get_color();
 
-	float red;
-	float green;
-	float blue;
+	double red;
+	double green;
+	double blue;
 	float threshold;
 	float slope;
 	int use_value;
